The great partition between MMS and Model Export -- Making Model Export tool independent of MMS
