+++ mms/model_service/mxnet_vision_service.py
@@ -11,30 +11,4 @@
 """`MXNetVisionService` defines a MXNet base vision service
 """
 
-from mms.model_service.mxnet_model_service import MXNetBaseService
-from mms.utils.mxnet import image, ndarray
-
-
-class MXNetVisionService(MXNetBaseService):
-    """MXNetVisionService defines a fundamental service for image classification task.
-    In preprocess, input image buffer is read to NDArray and resized respect to input
-    shape in signature.
-    In post process, top-5 labels are returned.
-    """
-    def _preprocess(self, data):
-        img_list = []
-        for idx, img in enumerate(data):
-            input_shape = self.signature['inputs'][idx]['data_shape']
-            # We are assuming input shape is NCHW
-            [h, w] = input_shape[2:]
-            img_arr = image.read(img)
-            img_arr = image.resize(img_arr, w, h)
-            img_arr = image.transform_shape(img_arr)
-            img_list.append(img_arr)
-        return img_list
-
-    def _postprocess(self, data):
-        assert hasattr(self, 'labels'), \
-            "Can't find labels attribute. Did you put synset.txt file into " \
-            "model archive or manually load class label file in __init__?"
-        return [ndarray.top_probability(d, self.labels, top=5) for d in data]
+from model_export.model_service.mxnet_vision_service import *
