+++ model_export/model_service/__init__.py
@@ -7,10 +7,10 @@
 # on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
-
 """
-MXNet Utils.
-
-Ideally this package should have completely been removed from the mms/utils/mxnet folder, but the existing model service codes rely on importing these utils through mms/utils/mxnet package.
-Hence to maintain backwards compatibility, we need to keep them here. The elegant part is that, the files in these packages just import * from the corresponding file in the model_export/utils/mxnet package
+Model services code
 """
+from . import model_service
+from . import mxnet_model_service
+from . import mxnet_vision_service
+from . import gluon_vision_service
