+++ model_export/model_service/mxnet_vision_service.py
@@ -1,0 +1,40 @@
+# Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Licensed under the Apache License, Version 2.0 (the "License").
+# You may not use this file except in compliance with the License.
+# A copy of the License is located at
+#     http://www.apache.org/licenses/LICENSE-2.0
+# or in the "license" file accompanying this file. This file is distributed
+# on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+# express or implied. See the License for the specific language governing
+# permissions and limitations under the License.
+
+"""`MXNetVisionService` defines a MXNet base vision service
+"""
+
+from model_export.model_service.mxnet_model_service import MXNetBaseService
+from model_export.utils.mxnet import image, ndarray
+
+
+class MXNetVisionService(MXNetBaseService):
+    """MXNetVisionService defines a fundamental service for image classification task.
+    In preprocess, input image buffer is read to NDArray and resized respect to input
+    shape in signature.
+    In post process, top-5 labels are returned.
+    """
+    def _preprocess(self, data):
+        img_list = []
+        for idx, img in enumerate(data):
+            input_shape = self.signature['inputs'][idx]['data_shape']
+            # We are assuming input shape is NCHW
+            [h, w] = input_shape[2:]
+            img_arr = image.read(img)
+            img_arr = image.resize(img_arr, w, h)
+            img_arr = image.transform_shape(img_arr)
+            img_list.append(img_arr)
+        return img_list
+
+    def _postprocess(self, data):
+        assert hasattr(self, 'labels'), \
+            "Can't find labels attribute. Did you put synset.txt file into " \
+            "model archive or manually load class label file in __init__?"
+        return [ndarray.top_probability(d, self.labels, top=5) for d in data]
