+++ model_export/utils/model_export_error_codes.py
@@ -9,8 +9,14 @@
 # permissions and limitations under the License.
 
 """
-MXNet Utils.
-
-Ideally this package should have completely been removed from the mms/utils/mxnet folder, but the existing model service codes rely on importing these utils through mms/utils/mxnet package.
-Hence to maintain backwards compatibility, we need to keep them here. The elegant part is that, the files in these packages just import * from the corresponding file in the model_export/utils/mxnet package
+This file defines all the errors generated by Model Export Tool
 """
+
+
+class ModelExportErrorCodes(object):
+    """
+    Class defining all the error codes generated by backend worker.
+    For explanation on each error please visit "docs/model_export_errors"
+    """
+    # Custom service error code
+    CUSTOM_SERVICE_ERROR = 12000
